src/Unet/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch
albumentations>=1.3.0
numpy
tqdm
Pillow
opencv-python


